+++ src/factory/unified-diff.cpp
@@ -6,164 +6,22 @@
 // Copyright: (c) 2024-2025 Anton van Wezenbeek
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <boost/tokenizer.hpp>
 #include <wex/core/log.h>
-#include <wex/core/regex.h>
-#include <wex/factory/frame.h>
 #include <wex/factory/unified-diff.h>
 
-#include <iostream>
+#include <algorithm>
 #include <utility>
 
-#define NEXT_TOKEN                                                             \
-  if (++tok_iter == tokens.end())                                              \
-  {                                                                            \
-    return false;                                                              \
-  }
-
-#define CHANGES_LINES(RANGE, TEXT)                                             \
-  for (int i = 0; i < m_range[RANGE]; i++)                                     \
-  {                                                                            \
-    NEXT_TOKEN                                                                 \
-    m_text[TEXT].push_back((*tok_iter).substr(1));                             \
-  }
-
-#define HEADER_LINES(REGEX, INTO)                                              \
-  if (!parse_header(REGEX, *tok_iter, INTO))                                   \
-  {                                                                            \
-    return false;                                                              \
-  }                                                                            \
-  NEXT_TOKEN
-
-#define SKIP_LINES                                                             \
-  while (tok_iter != tokens.end())                                             \
-  {                                                                            \
-    NEXT_TOKEN                                                                 \
-    if (!tok_iter->starts_with("diff ") && !tok_iter->starts_with("index "))   \
-    {                                                                          \
-      break;                                                                   \
-    }                                                                          \
-  }
-
-namespace wex
-{
-size_t stoi(const std::string& i)
-{
-  return i.empty() ? 1 : std::stoi(i);
-}
-} // namespace wex
+#include "unified-diff-parser.h"
 
 wex::factory::unified_diff::unified_diff(std::string input)
   : m_input(std::move(input))
 {
-  m_range.fill({0});
 }
 
 bool wex::factory::unified_diff::parse()
 {
-  using tokenizer = boost::tokenizer<boost::char_separator<char>>;
-
-  tokenizer tokens(m_input, boost::char_separator<char>("\r\n"));
-
-  tokenizer::iterator tok_iter = tokens.begin();
-
-  m_diffs = 0;
-  m_type  = diff_t::UNKNOWN;
-
-  while (tok_iter != tokens.end())
-  {
-    // skip first lines
-    SKIP_LINES;
-
-    // The unified output format starts with a two-line header
-    HEADER_LINES("--- a/(.*)", m_path[0]);
-    HEADER_LINES("\\+\\+\\+ b/(.*)", m_path[1]);
-
-    m_is_first = true;
-    m_is_last  = false;
-    m_type     = (m_type == diff_t::UNKNOWN ? diff_t::FIRST : diff_t::OTHER);
-
-    // Next come one or more hunks of differences
-    while (tok_iter != tokens.end())
-    {
-      regex r_hunk("@@ -([0-9]+),?([0-9]*) \\+([0-9]+),?([0-9]*) @@.*");
-
-      if (r_hunk.match(*tok_iter) != 4)
-      {
-        log("unified_diff") << *tok_iter << r_hunk.size();
-        return false;
-      }
-
-      m_range[0] = wex::stoi(r_hunk[0]);
-      m_range[1] = wex::stoi(r_hunk[1]);
-      m_range[2] = wex::stoi(r_hunk[2]);
-      m_range[3] = wex::stoi(r_hunk[3]);
-
-      m_text.fill({});
-
-      // Now get all - lines and all + lines, collect them, and invoke callback.
-      CHANGES_LINES(1, 0);
-      CHANGES_LINES(3, 1);
-
-      if (!report_diff())
-      {
-        return false;
-      }
-
-      trace("hunk");
-
-      m_diffs++;
-
-      m_is_first = false;
-      m_type     = diff_t::OTHER;
-
-      if (++tok_iter != tokens.end() && !(*tok_iter).starts_with("@@"))
-      {
-        m_is_last = true;
-
-        if (!report_diff())
-        {
-          return false;
-        }
-        trace("last");
-        m_diffs++;
-
-        break; // this was last hunk, continue with header lines
-      }
-    }
-  }
-
-  m_is_last = true;
-
-  if (m_type != diff_t::UNKNOWN)
-  {
-    m_type = diff_t::LAST;
-    m_diffs--;
-  }
-
-  report_diff_finish();
-
-  trace("finished");
-
-  return true;
-}
-
-bool wex::factory::unified_diff::parse_header(
-  const std::string& r,
-  const std::string& line,
-  path&              p)
-{
-  regex re(r);
-
-  if (!re.match(line))
-  {
-    log("unified_diff") << line << re.match_data().text();
-    return false;
-  }
-
-  p = path(re[0]);
-
-  return true;
+  return unified_diff_parser(this).parse();
 }
 
 void wex::factory::unified_diff::trace(const std::string& text) const